event-transformer: extract index selector
