+++ src/apps/selection/app.rs
@@ -130,7 +130,7 @@ impl App for Selection {
 
 #[cfg(test)]
 mod test {
-    use crate::midi::{Error, Event};
+    use crate::midi::Event;
     use crate::midi::features::{R, AppSelector};
     use crate::apps;
     use super::*;
@@ -147,10 +147,7 @@ mod test {
             return Ok(Event::SysEx(bytes));
         }
     }
-    impl EventTransformer for Transformer {
-        fn into_index(&self, _event: Event) -> Result<Option<u16>, Error> { Err(Error::Unsupported) }
-        fn from_index_to_highlight(&self, _index: u16) -> Result<Event, Error> { Err(Error::Unsupported) }
-    }
+    impl EventTransformer for Transformer {}
 
     const TRANSFORMER: Transformer = Transformer {};
