+++ src/midi/devices/default/mod.rs
@@ -1,4 +1,4 @@
-use crate::midi::{Error, Event, EventTransformer};
+use crate::midi::EventTransformer;
 
 pub fn transformer() -> &'static DefaultEventTransformer {
     return &DEFAULT_EVENT_TRANSFORMER;
@@ -6,18 +6,4 @@ pub fn transformer() -> &'static DefaultEventTransformer {
 
 const DEFAULT_EVENT_TRANSFORMER: DefaultEventTransformer = DefaultEventTransformer {};
 pub struct DefaultEventTransformer {}
-impl EventTransformer for DefaultEventTransformer {
-    fn into_index(&self, event: Event) -> Result<Option<u16>, Error> {
-         return match event {
-            // filter "note down" events, for notes higher than C2 (36), and with strictly positive velocity
-            Event::Midi([144, data1, data2, _]) if data1 >= 36 && data2 > 0 => {
-                Ok(Some((data1 - 36).into()))
-            },
-            _ => Ok(None),
-        };
-    }
-
-    fn from_index_to_highlight(&self, _index: u16) -> Result<Event, Error> {
-        return Err(Error::Unsupported);
-    }
-}
+impl EventTransformer for DefaultEventTransformer {}
