+++ src/midi/devices/launchpadpro/device.rs
@@ -1,6 +1,6 @@
 use std::convert::From;
 
-use crate::midi::{Reader, Writer, Error, Event, EventTransformer};
+use crate::midi::{Reader, Writer, Error, EventTransformer};
 
 pub struct LaunchpadPro<C> where C: Reader + Writer {
     pub connection: C,
@@ -41,12 +41,4 @@ impl LaunchpadProEventTransformer {
     }
 }
 
-impl EventTransformer for LaunchpadProEventTransformer {
-    fn into_index(&self, event: Event) -> Result<Option<u16>, Error> {
-        return super::index::into_index(event);
-    }
-
-    fn from_index_to_highlight(&self, index: u16) -> Result<Event, Error> {
-        return super::index::from_index_to_highlight(index);
-    }
-}
+impl EventTransformer for LaunchpadProEventTransformer {}
